requirements.txt
fastapi
uvicorn
pydantic-settings
langchain-groq
sentence-transformers
faiss-cpu
PyMuPDF
PyYAML
